refactor(http-client): use lambda for queryParams
This change modifies the `queryParams` in `ClientFeatures` to be a lambda expression. This allows for deferred evaluation of the query parameters, which can be useful in scenarios where the parameters need to be dynamically generated or fetched. It also modifies the usage to invoke the lambda to get the query params map.